Add Friend Layout 구현 - 파일입출력 #4
